RDKB-57687:[OneWifi] Add support for Operating Classes in dml subdoc
Reason for change: Operating classes as per the wfa data specs
is required for easy mesh configuration. This commit adds support for
that.
Along with this added the change to address mismatch in the channel value
of subdoc in Radioconfig and currentoperatingclass, give precedence
to channel in Radioconfig to ensure backward compatibility.

Test Procedure: 1) Compile change with halinterface change
		2) Run OneWifi and ensure dml subdoc has operating class
		   information added.
                3) Verified the fix working with channel change only in
                   Radioconfig.
Risks: Low
Priority: P1

Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>
Change-Id: I5d95e22bb9f639980582a483471420ad2b1515c5